refactor: product hooks를 model 계층으로 이동하여 FSD 주고 준수
